Cleanup of MEB routines, splitting code into MEB mMEB and MEBSandbox
