+++ dynamics/applications/benchmark_box_mevp.cpp
@@ -4,13 +4,13 @@
  * @author Thomas Richter <thomas.richter@ovgu.de>
  */
 
+#include "MEB.hpp"
 #include "Tools.hpp"
 #include "cgMomentum.hpp"
 #include "cgVector.hpp"
 #include "dgInitial.hpp"
 #include "dgTransport.hpp"
 #include "dgVisu.hpp"
-#include "meb.hpp"
 #include "mevp.hpp"
 #include "stopwatch.hpp"
