+++ dynamics/applications/compression_Vero_MEB_CGvel.cpp
@@ -4,13 +4,13 @@
  * @author Piotr Minakowski <piotr.minakowski@ovgu.no>
  */
 
+#include "MEB.hpp"
 #include "Tools.hpp"
 #include "cgMomentum.hpp"
 #include "cgVector.hpp"
 #include "dgInitial.hpp"
 #include "dgTransport.hpp"
 #include "dgVisu.hpp"
-#include "meb.hpp"
 #include "mevp.hpp"
 #include "stopwatch.hpp"
